# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface UserInfo {
  id: number;
  email: string;
}

export interface AuthResponse {
  token: string;
  user: UserInfo;
}

export interface ErrorResponse {
  error: string;
}